2024-2025-Spring/requirements.txt
cassandra-driver==3.29.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.1
geomet==0.2.1.post1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
